requirements.txt
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
icecream==2.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pyxlsb==1.0.10
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
streamlit==1.27.2
XlsxWriter==3.1.6
gspread==5.12.0
streamlit-authenticator==0.1.5